fix: consolidate getThemeStyles types and logic into a single file
